Changes to DiskFont class to use bitmap fonts with the advanceWidth included as a double precision float (8 bytes extra per char info block).
Tweaked so that the more likely accurate of the advanceWidth and bitmap width (in pixels) is used to advance the print location after each character has been printed. This gives rise to a 'tuning' value, which will be included in the config.csv file as a parameter.

TODO:
Include font tuning value in config.CSVFile.cpp
Migrate DiskFont code changes to the internal (rom) font code