Added support for font tuning variable in config.csv
Added support for disk fonts to be stored in SPIFFS memory on the ESP8266.

SPIFFS fonts are accessed transparently - if a font file having the same name and path exists on the SD card, it will be preferred, otherwise the SPIFFS filesystem will be checked, and if the font file named in the currently selected config.csv is found, it will be used.