Fixed reversion with font rendering.
Modified Config.cpp so that failure to get a time as well as an uncorrupted EEPROM config should trigger display of the image advising user to set the clock/eeprom config via the web interface.

Added bounds checking for input time value to Calendar.cpp, so that a negative value for the time will not be generated by calculations looking for Advent 1969 (the calendar epoch starts in 1970). Quick fix which may need to be improved was to make no year lower than 1971 acceptable as an input date.

Fixed bug getting lectionary settings from config.csv (clear config object every time a line of config.csv is scanned (if it is not the correct one whose number is stored in the EEPROM).

Fixed lone 'printf' in Temporale.cpp left from Windows.