Made changes to support diskfonts and the internal font having ascent, descent and lineheight metrics. The ascent metric is the size of the font in pixels above the line below which descenders are drawn.
The ascent metric is treated as the line at which a found is compared for y size bounds checking. In this cas, this means with the internal font that an extra line of text is drawn, above the number that would be drawn if the bottom of the bounding box of the font was used for comparison.