Added support for Daylight Savings Time. The support is rudimentary, given the complexity of the rules.
The start and end date of daylight savings is now automatically sent to the ESP8266 from the config.htm configuration page, and calculated using javascript in the browser. The timezone is no longer manually selectable - all values are populated in hidden form fields and sent automatically when the form is submitted.

Since DST start and end dates in some jurisdictions are not on fixed dates, but are on (e.g.) the last Sunday of March and October, DST observance will likely be off by a few days in these places in subsequent years. However, the start and end dates can be reset by using the configuration page again at least once per year.

Where fixed dates are used for daylight savings time start and end, this will not be necessary unless the rules change.