Modified startup to check battery status of DS3231 CR2032 battery. If the OSF (clock stopped) flag is set in the DS3231 clock chip, the battery is assumed to be failing and in need of being changed.
If the OSF flag is set, but the time has not been reset to the earliest possible values for the ds3231 (Jan 1 2000), then the image indicating that the battery should be changed is shown, since the time will be kept on very low power even if the oscillator has stopped (but it will not be updated).

If the time is reset as well, then it is assumed that the battery has been entirely removed (and replaced), so the image indicating that the time should be updated via the configuration webpage is shown in this case.