The clock now runs on the main battery, which will output between about 2.7v (empty) and 4.2v (full). The clock will run on the full range of these voltages, but the ESP8266 and peripherals will only run until the battery gets down to about 3.4 volts, when the 5v boost circuit will stop producing power for the 3.3v LDO regulator. So the clock should always run, even when the "charge battery" screen is displayed, and for quite some time after that.
Running the clock chip directly on the battery should not affect charging, since when 5v power is applied via usb, which is when the battery charges, the clock chip will switch over to the 3.3v LDO output supply automatically, even though it has a lower voltage than the battery (typically). See table 1, "power management" on page 10 of  the DS3231 data sheet for details of power management:

     SUPPLY CONDITION ACTIVE SUPPLY
     VCC < VPF, VCC < VBAT VBAT <- (VCC is less than power fail voltage (off), and VCC is less than vbat. This is the case when the 3.3V LDO output is off).
     VCC < VPF, VCC > VBAT VCC
     VCC > VPF, VCC < VBAT VCC  <- (VCC is on (3.3v), but VCC is less than VBAT - VCC will be 3.3v, vbat will be about 3.7v nominal, but the chip will still use VCC for power).
     VCC > VPF, VCC > VBAT VCC

Minor modifaction to the Config class, which checks that the DST settings are valid when getting the Local datetime - if not, returns the UTC datetime plus or minus the currently set timezone offset.

Also changed the Config class so that it is possible to use the debug=1|0 setting on its own - without any other parameters - when submitting the configuration query string to the url setconf.htm using the configuration webserver. This prevents the previous situation, where typically the time would be reset when the querystring previously used had the debug flag appended to it (this is currently the way debugging is turned on and off, but it could easily be included as a user interface element in the configuration web page).