Added optional verse numbering, newlines. Calibrated grey/red gradient of 4.2" epd against calculated, printed gradient.
Calibrated grey/red gradient of 4.2" epd against calculated, printed gradient.

Fixed bug which caused a space to appended to the wrong end of the numeric, ltr part of the references shown at the start of the reading when the reading text was rtl.

Added optional newline option after each verse, and also option for superscript verse numbers at the beginning of each verse.