Added enclosure design
Added a prototype enclosure for the 4.2 in Lectionary.

JLCPCB printed the current boards 102mm wide instead of 100mm - this design is for the 102mm board.

A further iteration of the board which should hopefully be 100mm wide is in design, and when this board is printed I will also upload a revised case design.