Fixed wrapping on full stops followed by a single or double quote
Corrected display bugs: a single or double quote after a full stop could get wrapped, this is now not allowed - wrapping will occur after a quote if it follows a fullstop, rather than at the full stop.

Found and hopefully fixed what appeared to be a bug in the ellipsis case processing.

Spaces are now added to the end of verses which do not have them, if the verses are displayed without linebreaks.