Added 1 byte, 2 pixel cache to Framebuffer code
Added 1 byte, 2 pixel cache to Framebuffer code, which gave about 15% speed improvement in rendering in a typical test. The pixels are encoded in the framebuffer SPI RAM chip as 4bits per pixel, 2 pixels per byte, which was causing thrashing of the SPI bus. This solution doesn't eliminate that, but does improve it.