Version 2.5 enclosure with glass bezel
Version 2.5 enclosure with glass bezel; minor fix to romancalendar.ino to handle case when Lectionary is left on USB power (charging) for more than one hour - screen used to update every hour in this case, whether it needed to or not. Now it will update at the same frequency as when on battery power.