Make sure WiFi circuit is off when on battery to save power
Discovered WiFi circuit may be powered on by default after boot, even when not connected to a network. Forced it off if on battery, and if power is removed after the webserver is run - this should save power. Figures from the internet suggest ESP8266 goes from about 75mA to 25mA when Wifi circuit is off. (Will test)